requirements.txt
apiflask==2.3.2
Flask==3.1.0
mysql-connector-python==9.2.0
flask-sqlalchemy==3.1.1
Flask-migrate==4.1.0
flask-cors==5.0.1
pyJWT==2.10.1
requests==2.31.0
python-dotenv==1.0.1